feat: include representation to definition form
